Micronaut code completion does not implement CompletionItem.
