+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/completion/MicronautConfigCompletionProvider.java
@@ -18,20 +18,33 @@
  */
 package org.netbeans.modules.micronaut.completion;
 
+import java.awt.Color;
+import java.util.regex.Pattern;
+import javax.swing.text.BadLocationException;
 import javax.swing.text.Document;
 import javax.swing.text.JTextComponent;
 import org.netbeans.api.editor.document.EditorDocumentUtils;
+import org.netbeans.api.editor.document.LineDocument;
+import org.netbeans.api.editor.document.LineDocumentUtils;
 import org.netbeans.api.editor.mimelookup.MimeRegistration;
 import org.netbeans.api.project.FileOwnerQuery;
 import org.netbeans.api.project.Project;
+import org.netbeans.lib.editor.codetemplates.api.CodeTemplateManager;
+import org.netbeans.lib.editor.util.ArrayUtilities;
+import org.netbeans.modules.editor.indent.api.IndentUtils;
 import org.netbeans.modules.micronaut.MicronautConfigProperties;
 import org.netbeans.modules.micronaut.MicronautConfigUtilities;
+import org.netbeans.spi.editor.completion.CompletionItem;
 import org.netbeans.spi.editor.completion.CompletionProvider;
 import org.netbeans.spi.editor.completion.CompletionResultSet;
 import org.netbeans.spi.editor.completion.CompletionTask;
 import org.netbeans.spi.editor.completion.support.AsyncCompletionQuery;
 import org.netbeans.spi.editor.completion.support.AsyncCompletionTask;
+import org.netbeans.spi.editor.completion.support.CompletionUtilities;
+import org.netbeans.swing.plaf.LFCustoms;
 import org.openide.filesystems.FileObject;
+import org.openide.util.Exceptions;
+import org.openide.xml.XMLUtil;
 import org.springframework.boot.configurationmetadata.ConfigurationMetadataProperty;
 
 /**
@@ -39,6 +52,8 @@ import org.springframework.boot.configurationmetadata.ConfigurationMetadataPrope
  * @author Dusan Balek
  */
 public class MicronautConfigCompletionProvider implements CompletionProvider {
+
+    public static final String PROPERTY_NAME_COLOR = getHTMLColor(64, 64, 217);
 
     @MimeRegistration(mimeType = "text/x-yaml", service = CompletionProvider.class)
     public static MicronautConfigCompletionProvider createYamlProvider() {
@@ -75,11 +90,28 @@ public class MicronautConfigCompletionProvider implements CompletionProvider {
         return 0;
     }
 
-    static CompletionTask createDocTask(ConfigurationMetadataProperty element) {
-        return new AsyncCompletionTask(new MicronautConfigDocumentationQuery(element));
+    private static String getHTMLColor(int r, int g, int b) {
+        Color c = LFCustoms.shiftColor(new Color(r, g, b));
+        return "<font color=#" //NOI18N
+                + LFCustoms.getHexString(c.getRed())
+                + LFCustoms.getHexString(c.getGreen())
+                + LFCustoms.getHexString(c.getBlue())
+                + ">"; //NOI18N
+    }
+
+    private static String escape(String s) {
+        if (s != null) {
+            try {
+                return XMLUtil.toAttributeValue(s);
+            } catch (Exception ex) {}
+        }
+        return s;
     }
 
     private static class MicronautConfigCompletionQuery extends AsyncCompletionQuery {
+
+        private static final String ICON = "org/netbeans/modules/editor/resources/completion/field_16.png"; //NOI18N
+        private static final Pattern FQN = Pattern.compile("(\\w+\\.)+(\\w+)");
 
         private final Project project;
 
@@ -89,17 +121,133 @@ public class MicronautConfigCompletionProvider implements CompletionProvider {
 
         @Override
         protected void query(CompletionResultSet resultSet, Document doc, int caretOffset) {
-            resultSet.addAllItems(new MicronautConfigCompletionTask().query(doc, caretOffset, project, new MicronautConfigCompletionTask.ItemFactory<MicronautConfigCompletionItem>() {
+            resultSet.addAllItems(new MicronautConfigCompletionTask().query(doc, caretOffset, project, new MicronautConfigCompletionTask.ItemFactory<CompletionItem>() {
                 @Override
-                public MicronautConfigCompletionItem createPropertyItem(ConfigurationMetadataProperty property, int offset, int baseIndent, int indentLevelSize, int idx) {
+                public CompletionItem createPropertyItem(ConfigurationMetadataProperty property, int offset, int baseIndent, int indentLevelSize, int idx) {
                     resultSet.setAnchorOffset(offset);
-                    return MicronautConfigCompletionItem.createPropertyItem(property, offset, baseIndent, indentLevelSize, idx);
+                    String propName = property.getId();
+                    String propType = property.getType();
+                    CompletionUtilities.CompletionItemBuilder builder = CompletionUtilities.newCompletionItemBuilder(propName)
+                            .iconResource(ICON)
+                            .leftHtmlText(property.isDeprecated()
+                                    ? PROPERTY_NAME_COLOR + "<s>" + propName + "</s></font>"
+                                    : PROPERTY_NAME_COLOR + propName + "</font>")
+                            .sortPriority(property.isDeprecated() ? 30 : 20)
+                            .documentationTask(() -> {
+                                return new AsyncCompletionTask(new MicronautConfigDocumentationQuery(property));
+                            })
+                            .onSelect((component, overwrite) -> {
+                                try {
+                                    Document doc = component.getDocument();
+                                    LineDocument lineDocument = LineDocumentUtils.as(doc, LineDocument.class);
+                                    if (lineDocument != null) {
+                                        int caretOffset = component.getCaretPosition();
+                                        int end = LineDocumentUtils.getWordEnd(lineDocument, caretOffset);
+                                        if (overwrite && LineDocumentUtils.getWordStart(lineDocument, end) == offset) {
+                                            String textEnd = doc.getText(end, 1);
+                                            while(baseIndent < 0 && textEnd.endsWith(".")) {
+                                                end = LineDocumentUtils.getWordEnd(lineDocument, end + 1);
+                                                textEnd = doc.getText(end, 1);
+                                            }
+                                            if (baseIndent < 0 && textEnd.endsWith("=") || textEnd.endsWith(":")) {
+                                                end++;
+                                            }
+                                            doc.remove(offset, Math.max(caretOffset, end) - offset);
+                                        } else if (offset < caretOffset) {
+                                            doc.remove(offset, caretOffset - offset);
+                                        }
+                                        StringBuilder sb = new StringBuilder();
+                                        String name = propName.substring(idx);
+                                        String[] parts = name.split("\\.");
+                                        if (baseIndent < 0) {
+                                            int num = 1;
+                                            for (int i = 0; i < parts.length; i++) {
+                                                String part = parts[i];
+                                                if ("*".equals(part)) {
+                                                    sb.append("${PAR#" + num++ + " default=\"\"}");
+                                                } else {
+                                                    sb.append(part);
+                                                }
+                                                if (i < parts.length - 1) {
+                                                    sb.append(".");
+                                                } else {
+                                                    sb.append("=${cursor}");
+                                                }
+                                            }
+                                        } else {
+                                            int lineStart = LineDocumentUtils.getLineStart(lineDocument, caretOffset);
+                                            int lineIndent = IndentUtils.lineIndent(doc, lineStart);
+                                            ArrayUtilities.appendSpaces(sb, baseIndent - lineIndent);
+                                            int indent = baseIndent;
+                                            int num = 1;
+                                            for (int i = 0; i < parts.length; i++) {
+                                                String part = parts[i];
+                                                if ("*".equals(part)) {
+                                                    sb.append("${PAR#" + num++ + " default=\"\"}");
+                                                } else {
+                                                    sb.append(part);
+                                                }
+                                                if (i < parts.length - 1) {
+                                                    sb.append(":\n");
+                                                    ArrayUtilities.appendSpaces(sb, (indent = indent + indentLevelSize));
+                                                } else {
+                                                    sb.append(": ${cursor}");
+                                                }
+                                            }
+                                        }
+                                        CodeTemplateManager.get(doc).createTemporary(sb.toString()).insert(component);
+                                    }
+                                } catch (BadLocationException ex) {
+                                    Exceptions.printStackTrace(ex);
+                                }
+                            });
+                    if (propType != null) {
+                        builder.rightHtmlText(escape(FQN.matcher(propType).replaceAll("$2")));
+                    }
+                    return builder.build();
                 }
 
                 @Override
-                public MicronautConfigCompletionItem createTopLevelPropertyItem(String propName, int offset, int baseIndent, int indentLevelSize) {
+                public CompletionItem createTopLevelPropertyItem(String propName, int offset, int baseIndent, int indentLevelSize) {
                     resultSet.setAnchorOffset(offset);
-                    return MicronautConfigCompletionItem.createTopLevelPropertyItem(propName, offset, baseIndent, indentLevelSize);
+                    return CompletionUtilities.newCompletionItemBuilder(propName)
+                            .iconResource(ICON)
+                            .leftHtmlText(PROPERTY_NAME_COLOR + "<b>" + propName + "</b></font>")
+                            .sortPriority(10)
+                            .onSelect((component, overwrite) -> {
+                                try {
+                                    Document doc = component.getDocument();
+                                    LineDocument lineDocument = LineDocumentUtils.as(doc, LineDocument.class);
+                                    if (lineDocument != null) {
+                                        int caretOffset = component.getCaretPosition();
+                                        int end = LineDocumentUtils.getWordEnd(lineDocument, caretOffset);
+                                        if (overwrite && LineDocumentUtils.getWordStart(lineDocument, end) == offset) {
+                                            String textEnd = doc.getText(end, 1);
+                                            if (baseIndent < 0 && textEnd.endsWith(".") || textEnd.endsWith(":")) {
+                                                end++;
+                                            }
+                                            doc.remove(offset, Math.max(caretOffset, end) - offset);
+                                        } else if (offset < caretOffset) {
+                                            doc.remove(offset, caretOffset - offset);
+                                        }
+                                        StringBuilder sb = new StringBuilder();
+                                        if (baseIndent < 0) {
+                                            sb.append("*".equals(propName) ? "${PAR#1 default=\"\"}" : propName).append(".${cursor completionInvoke}");
+                                        } else {
+                                            int lineStart = LineDocumentUtils.getLineStart(lineDocument, caretOffset);
+                                            int lineIndent = IndentUtils.lineIndent(doc, lineStart);
+                                            ArrayUtilities.appendSpaces(sb, baseIndent - lineIndent);
+                                            sb.append("*".equals(propName) ? "${PAR#1 default=\"\"}" : propName).append(":\n");
+                                            ArrayUtilities.appendSpaces(sb, baseIndent + indentLevelSize);
+                                            sb.append("${cursor completionInvoke}");
+                                        }
+                                        CodeTemplateManager.get(doc).createTemporary(sb.toString()).insert(component);
+                                    }
+                                } catch (BadLocationException ex) {
+                                    Exceptions.printStackTrace(ex);
+                                }
+                            })
+                            .build();
                 }
             }));
             resultSet.finish();
